more stuff
tweaks patrol priority & distance
tweaks resupply time
fixes flak runtime
de-fractures structure/overmap/Initialize()
Adds failsafe for weird obj_integrity vs max_integrity inits
Adds failsafe for the has gravity check
Adds integrity values to a bunch of ships to prevent issues.
Changes syndi bombers to be Destroyer-swarmers instead of just Destroyers

+++ nsv13/code/modules/munitions/ship_weapons/ballistic_weapons/flak.dm
@@ -157,9 +157,14 @@
 
 /obj/item/projectile/guided_munition/Crossed(atom/movable/AM) //Here, we check if the bullet that hit us is from a friendly ship. If it's from an enemy ship, we explode as we've been flak'd down.
 	. = ..()
+
+	if(!isprojectile(AM))
+		return
+
 	var/obj/item/projectile/P = AM //This is hacky, refactor check_faction to unify both of these. I'm bodging it for now.
 	if(P.damage <= 0)
 		return
+
 	if(isprojectile(AM) && P.faction != faction) //Because we could be in the same faction and collide with another bullet. Let's not blow ourselves up ok?
 		if(obj_integrity <= P.damage) //Tank the hit, take some damage
 			qdel(P)
